pyproject.toml
[project]
name = "ai-telegram-bot"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "nikmosi", email = "i@nikmosi.ru"},
    {name = "*8_a.k_8", email = "kinglook1998@gmail.com"}
]
dependencies = [
    "aiogram>=3.13.1",
    "g4f>=0.3.3.1",
    "pydantic-settings>=2.6.0",
    "aiohttp-socks>=0.9.0",
    "loguru>=0.7.2",
    "SpeechRecognition>=3.11.0",
    "ffmpeg-python>=0.2.0",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[project.optional-dependencies]
dev = [
    "pytest>=8.3.3",
    "pytest-timeout>=2.3.1",
    "mypy>=1.13.0",
    "loguru-mypy>=0.0.4",
]
[tool.pdm]
distribution = false

[tool.pdm.scripts]
test = "pytest"
lint = "pre-commit run --all-files"
mypy = "mypy --strict ."
up = "docker compose up --force-recreate --build"

[tool.mypy]
explicit_package_bases = true
mypy_path = "src"
ignore_missing_imports = true
python_version = "3.12"
strict = true
namespace_packages = true
exclude = "tests/.*"
[tool.pytest.ini_options]
pythonpath = [
  "src"
]
